requirements.txt
google-cloud-translate
grpcio
grpcio-tools
redis
pytest